addon: remove authentication package Since moving to a single authentication method the authentication package does not make sense. This commit removes it and moves the remaining function to the addon package.
Since we also now only provide one authentication method there is no
more need to use annotations on the root resources
