+++ hack/addon-install/templates/clf-instance.yaml
@@ -3,8 +3,6 @@ kind: ClusterLogForwarder
 metadata:
   name: instance
   namespace: open-cluster-management
-  annotations:
-    authentication.mcoa.openshift.io/cw: SecretReference
 spec:
   outputs:
    - cloudwatch:
