+++ hack/addon-install/templates/instance-default.yaml
@@ -3,8 +3,6 @@ kind: ClusterLogForwarder
 metadata:
   name: instance-default
   namespace: open-cluster-management
-  annotations:
-    authentication.mcoa.openshift.io/cw: SecretReference
 spec:
   outputs:
    - cloudwatch:
