+++ internal/logging/handlers/handler.go
@@ -5,7 +5,6 @@ import (
 
 	loggingv1 "github.com/openshift/cluster-logging-operator/apis/logging/v1"
 	"github.com/rhobs/multicluster-observability-addon/internal/addon"
-	"github.com/rhobs/multicluster-observability-addon/internal/addon/authentication"
 	"github.com/rhobs/multicluster-observability-addon/internal/logging/manifests"
 	addonapiv1alpha1 "open-cluster-management.io/api/addon/v1alpha1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -27,22 +26,17 @@ func BuildOptions(k8s client.Client, mcAddon *addonapiv1alpha1.ManagedClusterAdd
 	}
 	resources.ClusterLogForwarder = clf
 
-	targetSecretName := make(map[authentication.Target]string)
+	targetSecretName := make(map[addon.Target]string)
 	for _, output := range clf.Spec.Outputs {
-		targetSecretName[authentication.Target(output.Name)] = output.Secret.Name
+		targetSecretName[addon.Target(output.Name)] = output.Secret.Name
 	}
 
 	ctx := context.Background()
-	secretsProvider := authentication.NewSecretsProvider(k8s, clf.Namespace, mcAddon.Namespace)
-	targetSecrets, err := secretsProvider.GenerateSecrets(ctx, clf.Annotations, targetSecretName)
+	targetSecrets, err := addon.GetSecrets(ctx, k8s, clf.Namespace, mcAddon.Namespace, targetSecretName)
 	if err != nil {
 		return resources, err
 	}
-
-	resources.Secrets, err = secretsProvider.FetchSecrets(ctx, targetSecrets)
-	if err != nil {
-		return resources, err
-	}
+	resources.Secrets = targetSecrets
 
 	return resources, nil
 }
