+++ internal/tracing/handlers/handler.go
@@ -7,7 +7,6 @@ import (
 
 	otelv1alpha1 "github.com/open-telemetry/opentelemetry-operator/apis/v1alpha1"
 	"github.com/rhobs/multicluster-observability-addon/internal/addon"
-	"github.com/rhobs/multicluster-observability-addon/internal/addon/authentication"
 	"github.com/rhobs/multicluster-observability-addon/internal/tracing/manifests"
 	"github.com/rhobs/multicluster-observability-addon/internal/tracing/manifests/otelcol"
 	v1 "k8s.io/api/core/v1"
@@ -46,22 +45,17 @@ func BuildOptions(k8s client.Client, mcAddon *addonapiv1alpha1.ManagedClusterAdd
 	}
 
 	ctx := context.Background()
-	secretsProvider := authentication.NewSecretsProvider(k8s, otelCol.Namespace, mcAddon.Namespace)
-	targetsSecret, err := secretsProvider.GenerateSecrets(ctx, otelCol.Annotations, targetSecretName)
+	targetSecrets, err := addon.GetSecrets(ctx, k8s, otelCol.Namespace, mcAddon.Namespace, targetSecretName)
 	if err != nil {
 		return resources, err
 	}
-
-	resources.Secrets, err = secretsProvider.FetchSecrets(ctx, targetsSecret)
-	if err != nil {
-		return resources, err
-	}
+	resources.Secrets = targetSecrets
 
 	return resources, nil
 }
 
-func buildExportersSecrets(otelCol *otelv1alpha1.OpenTelemetryCollector) (map[authentication.Target]string, error) {
-	exporterSecrets := map[authentication.Target]string{}
+func buildExportersSecrets(otelCol *otelv1alpha1.OpenTelemetryCollector) (map[addon.Target]string, error) {
+	exporterSecrets := map[addon.Target]string{}
 
 	exporters, err := getExporters(otelCol)
 	if err != nil {
@@ -85,7 +79,7 @@ func buildExportersSecrets(otelCol *otelv1alpha1.OpenTelemetryCollector) (map[au
 				continue
 			}
 			klog.Info("exporter ", exporter, " uses secret ", vol.Secret.SecretName)
-			exporterSecrets[authentication.Target(exporter)] = vol.Secret.SecretName
+			exporterSecrets[addon.Target(exporter)] = vol.Secret.SecretName
 		}
 	}
 	return exporterSecrets, nil
