+++ internal/tracing/manifests/options.go
@@ -2,14 +2,14 @@ package manifests
 
 import (
 	otelv1alpha1 "github.com/open-telemetry/opentelemetry-operator/apis/v1alpha1"
-	"github.com/rhobs/multicluster-observability-addon/internal/addon/authentication"
+	"github.com/rhobs/multicluster-observability-addon/internal/addon"
 	corev1 "k8s.io/api/core/v1"
 	addonapiv1alpha1 "open-cluster-management.io/api/addon/v1alpha1"
 )
 
 type Options struct {
 	ClusterName            string
-	Secrets                map[authentication.Target]corev1.Secret
+	Secrets                map[addon.Target]corev1.Secret
 	ConfigMaps             []corev1.ConfigMap
 	OpenTelemetryCollector *otelv1alpha1.OpenTelemetryCollector
 	AddOnDeploymentConfig  *addonapiv1alpha1.AddOnDeploymentConfig
