+++ internal/tracing/manifests/otelcol/volumemount_test.go
@@ -14,9 +14,6 @@ func Test_ConfigureVolumeMounts(t *testing.T) {
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      "tracing-otlphttp-auth",
 			Namespace: "cluster-1",
-			Annotations: map[string]string{
-				annotation: "otlphttp",
-			},
 		},
 		Data: map[string][]byte{
 			"tls.crt": []byte("data"),
